--- examples/zombie_session_test/kibana.jsonc
@@ -1,13 +1,0 @@
-{
-  "type": "plugin",
-  "id": "@kbn/zombie-session-test-plugin",
-  "owner": "@elastic/kibana-core",
-  "description": "Test plugin to reproduce HTTP/2 zombie sessions by toggling iptables rules on ES connections.",
-  "group": "platform",
-  "visibility": "private",
-  "plugin": {
-    "id": "zombieSessionTest",
-    "server": true,
-    "browser": false
-  }
-}
