--- examples/zombie_session_test/moon.yml
@@ -1,31 +1,0 @@
-# This file is generated by the @kbn/moon package. Any manual edits will be erased!
-#  To extend this, write your extensions/overrides to 'moon.extend.yml'
-#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/zombie-session-test-plugin'
-
-$schema: https://moonrepo.dev/schemas/project.json
-id: '@kbn/zombie-session-test-plugin'
-layer: unknown
-owners:
-  defaultOwner: '@elastic/kibana-core'
-toolchains:
-  default: node
-language: typescript
-project:
-  title: '@kbn/zombie-session-test-plugin'
-  description: Moon project for @kbn/zombie-session-test-plugin
-  channel: ''
-  owner: '@elastic/kibana-core'
-  sourceRoot: examples/zombie_session_test
-dependsOn:
-  - '@kbn/core'
-  - '@kbn/config-schema'
-tags:
-  - plugin
-  - prod
-  - group-platform
-  - private
-fileGroups:
-  src:
-    - server/**/*.ts
-    - '!target/**/*'
-tasks: {}
