--- examples/zombie_session_test/server/index.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import type { PluginInitializer } from '@kbn/core/server';
-
-export const plugin: PluginInitializer<{}, {}> = async () => {
-  const { ZombieSessionTestPlugin } = await import('./plugin');
-  return new ZombieSessionTestPlugin();
-};
