--- examples/zombie_session_test/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "@kbn/tsconfig-base/tsconfig.json",
-  "compilerOptions": {
-    "outDir": "target/types"
-  },
-  "include": [
-    "server/**/*.ts"
-  ],
-  "exclude": [
-    "target/**/*"
-  ],
-  "kbn_references": [
-    "@kbn/core",
-    "@kbn/config-schema"
-  ]
-}
