+++ src/platform/plugins/shared/dashboard/moon.yml
@@ -121,6 +121,7 @@ dependsOn:
   - '@kbn/monaco'
   - '@kbn/code-editor'
   - '@kbn/as-code-shared-schemas'
+  - '@kbn/as-code-shared-transforms'
   - '@kbn/core-http-common'
   - '@kbn/core-http-server-mocks'
   - '@kbn/core-http-browser'
