+++ src/platform/plugins/shared/discover/moon.yml
@@ -133,6 +133,7 @@ dependsOn:
   - '@kbn/as-code-filters-schema'
   - '@kbn/as-code-filters-transforms'
   - '@kbn/as-code-shared-schemas'
+  - '@kbn/as-code-shared-transforms'
   - '@kbn/scout'
   - '@kbn/scout-synthtrace'
   - '@kbn/synthtrace-client'
