--- tmp/zombie_session_tests/docker/certs/kibana.crt
@@ -1,29 +1,0 @@
-Bag Attributes
-    friendlyName: kibana
-    localKeyID: 54 69 6D 65 20 31 37 37 31 34 33 37 30 36 30 33 34 31
-subject=CN=kibana
-issuer=CN=Elastic Certificate Tool Autogenerated CA
------BEGIN CERTIFICATE-----
-MIIEBTCCAu2gAwIBAgIVAKqEEFDGSEs427xhISo7Je1FMv9rMA0GCSqGSIb3DQEB
-CwUAMDQxMjAwBgNVBAMTKUVsYXN0aWMgQ2VydGlmaWNhdGUgVG9vbCBBdXRvZ2Vu
-ZXJhdGVkIENBMCAXDTI2MDIxODE3NTEwMFoYDzIwNzYwMjA2MTc1MTAwWjARMQ8w
-DQYDVQQDEwZraWJhbmEwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDn
-WglNV6V9dB8F27upvq8rUT5hbhNrucawkvOJrcH0FP2pvinqHCboWRxMPO8P13Yd
-xJ/3oD0gLQZYVF9Jp3jduU9Xss5p/J0MeyhB8p/DPwGIH8mWeDTWgmVGJqkmi0+x
-wm4Me5h6W0OXjfuigS1pLRBM3LhfMSnYGuL5c5cVkFxSk/h7mjB+nPTMyqsa5sIc
-/pZdJ+yELxnPXM27HaTutvzKc9qmq7aLk+4z6aHx/stTJIwMaavhqEX62xCFhZZV
-xKLu8kc32X25lsym92E1eDviCeKcv9319TWe9CFHit40xOJHrLTirWDqrfVTXbIE
-Fvus0UzM+qshEJt7iPQzAgMBAAGjggEtMIIBKTAdBgNVHQ4EFgQUOZVqGbUpsDWb
-zODy6rn0QlE5F10wHwYDVR0jBBgwFoAUMEwqwI5b0MYpNxwaHJ9Tw1Lp3p4wgdsG
-A1UdEQSB0zCB0IIEZXMwMYI8a2Ita2ItdWlhbS5wcm9qZWN0LWFiY2RlZjEyMzQ1
-Njc4OTAxMjM0NTY3ODkwMTIzNDU2LmtiLmxvY2FsggRlczAygl1FSURFTlQuYWJj
-ZGVmMTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTYucHJvamVjdC5lbGFzdGljc2Vh
-cmNoLnR5cGUub3JnMTIzNDU2Nzg5MC5vcmdhbml6YXRpb26CBGVzMDOCCWxvY2Fs
-aG9zdIIUaG9zdC5kb2NrZXIuaW50ZXJuYWwwCQYDVR0TBAIwADANBgkqhkiG9w0B
-AQsFAAOCAQEABRwp6ooJzO/nzBO/J+GBV0+brtTgqSXg8hZ6+9BUCv2dNPV7U/io
-TkZtY3mbK2cgq0B6Q2e/poiZtHW35Zuo87zYonSyQxqblML7C8lzFEM4qTAOzBAN
-xzzbSciumz5p88x5ug/y+0VDB2BkyQ+24Pw2YxbuA7iHjOOmdcgScPfulAx4FImf
-p2AZ/DeRdArTZAyI2pTKkOxvNRtToiBII2DdmHe7oYcfXFG02Z9vwz/o4YO+cLTs
-8hMMrL9G1y56wsSWT+kh/j3QDzqMnNmQ+VeEjEvxET6TbHZqgRa4q2qkyS2ls+X8
-NHWZ5uNXYbfnZpDdZ7LyAvvNvkThWc87gw==
------END CERTIFICATE-----
