--- tmp/zombie_session_tests/docker/certs/kibana.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEA51oJTVelfXQfBdu7qb6vK1E+YW4Ta7nGsJLzia3B9BT9qb4p
-6hwm6FkcTDzvD9d2HcSf96A9IC0GWFRfSad43blPV7LOafydDHsoQfKfwz8BiB/J
-lng01oJlRiapJotPscJuDHuYeltDl437ooEtaS0QTNy4XzEp2Bri+XOXFZBcUpP4
-e5owfpz0zMqrGubCHP6WXSfshC8Zz1zNux2k7rb8ynPapqu2i5PuM+mh8f7LUySM
-DGmr4ahF+tsQhYWWVcSi7vJHN9l9uZbMpvdhNXg74gninL/d9fU1nvQhR4reNMTi
-R6y04q1g6q31U12yBBb7rNFMzPqrIRCbe4j0MwIDAQABAoIBAEfBllcF0094/9JK
-WAeLHDu8RDPl92IXSfgkbCMM4LZ6+D54u/lf5/VzBiLjV2a7dmg/PzqH2c2sCWbG
-LJPkvMlQm7pgvEMFVhSz8kWaIlFlrmzxJw5jlEfgQ9chUJ+i4AILgySeBSoWOn28
-TQsXM+WGU6LzZsAnuInNtQ6X+ol2GQklRQ2e/y9TI69WQ4OPaTP4VbyXnXsMrUoE
-v6hE3AKwcXUkaZ2l+zZGS5CJ8AuVbOLYacJZUXIl6V5FYL3FxR7H4rieR0RhG2PY
-RQv7NpCZmicSP3rTQ3IUlbq7I5G222+XtO4L5sOIuNA7+QsILS2i/t34L2kBovQd
-mwWE+ekCgYEA/NAZ2XCEf3eqjxIyr93aQiVUhVIN6Q4sUQT2pzagVpb4k7j7/QId
-ME8w7T2xXZuoJxGTZHKl42uwlMjN7e2rYVLUFbvABS9h11t2+boYsF+n9FjYmh2L
-F6j5yN6/u7ibOSZSQGENMowZ8/t4DXNXs6htwemc+rtm7OcONa0z2EUCgYEA6kSs
-i6S7YbsQO47ROTDlf7xnAbrZAH41KAslDkLXs+6tnuNTO45e+8K764dcPasdrDMX
-twMzlVXmBCNLUxdSwk9pLxpse1pjH9nQq+SeE4MhnXto12Nzt2kA6xDFTO0Ywzav
-ILn3cYUEShMUytu2pe/0R7EoCVxPUeU4C/iFzhcCgYEAxmrdnta8Zv6YkmmJ9pV8
-c2WxDGH2IO/KwFvQ6jPpa4xZ1DbfLxe1qPC+SbSdvAYq117n+3Iv2Gnw0RU46oAa
-fevwII6WintBozBaFG2GawboXtJMTcjaHdu1D34jpUWiLhxxea2yGfXzeJXpB0V7
-k7mhSwv69J6YjV5avK4PfrECgYEAg/eHk2qKu/UaodJD/gmTXq+M/yZ4U5TE8PfG
-OhBhXlTXrSe1nVkIHJ6IKZeo2HxqTLTDaS7+geNPnYkcR4Rd9GOzhvtFnP8/05Np
-v0sb2TYHW6VHW/4EE4+tGr3pxvnQ9zb41GCuCV67GddB4Tx/2V4gp7oeKZe8fw+2
-0NeA2KUCgYEAwMqYbBi+5RpOhmImd+gHK2bMzJbxR6HHjwZ1BbmcCyrOfWsC6p16
-wgJTbgZV2LbBO9OtPQzbSKTnPoiRUiodRtPOikzkCyDN0FfD/SagCMu2fdlkOhHF
-f17UItNbc0prYHeBC47KlanKE6/CkbwRmBYpWrx4wis2BHzoifIXIfU=
------END RSA PRIVATE KEY-----
