--- tmp/zombie_session_tests/kibana-zombie/zombie_session_test/kibana.json
@@ -1,10 +1,0 @@
-{
-  "id": "zombieSessionTest",
-  "version": "kibana",
-  "server": true,
-  "ui": false,
-  "owner": {
-    "name": "@elastic/kibana-core",
-    "githubTeam": "kibana-core"
-  }
-}
