--- tmp/zombie_session_tests/macos/logs.txt
@@ -1,88 +1,0 @@
-
-======================================================================
- HTTP/2 Zombie Session Reproduction (macOS)
- Node.js: v22.22.0 | PID: 10210
- Mode: CRASH (will trigger assertion failure)
-======================================================================
-
-[16:19:58.929] [PHASE 1] Starting HTTP/2 server...
-[16:19:58.979] [SERVER] Listening on https://127.0.0.1:18443
-[16:19:59.481] [PHASE 2] Connecting HTTP/2 client...
-[16:19:59.507] [PHASE 2] Connected. Verifying with initial request...
-[16:19:59.514] [PHASE 2] Initial request succeeded: "ok from server: /health-check"
-[16:19:59.514] [PHASE 3] Activating firewall black hole...
-[16:19:59.514] [PHASE 3]   Rule: block drop all TCP to/from 127.0.0.1:18443
-[16:19:59.514] [PHASE 3]   Effect: packets silently disappear (no RST, no FIN, no ICMP)
-[16:19:59.514] [PHASE 3]   This simulates: NAT timeout, network partition, peer crash
-[16:19:59.590] [PHASE 3] Black hole ACTIVE. Connection is now silently broken.
-
-[16:19:59.590] [PHASE 4] Sending HTTP/2 requests into the black hole...
-[16:19:59.590] [PHASE 4] (These requests will never reach the server)
-[16:19:59.590] [PHASE 4]   Sent request 1/5
-[16:19:59.892] [PHASE 4]   Sent request 2/5
-[16:20:00.195] [PHASE 4]   Sent request 3/5
-[16:20:00.497] [PHASE 4]   Sent request 4/5
-[16:20:00.799] [PHASE 4]   Sent request 5/5
-
-[16:20:01.102] [PHASE 5] Monitoring zombie session state...
-[16:20:01.102] [PHASE 5] Watch: session thinks it's healthy, but writes are stuck
-
-  ┌──────────────────────────────────────────────────────────────────────────
-  │ [1.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [2.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [3.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │         ping=sent (callback pending...)
-  │ [4.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [5.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [6.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [7.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [8.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [9.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [10.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [11.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [12.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [13.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [14.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  │ [15.0s] ZOMBIE | closed=false destroyed=false responses=0/5 events=none
-  └──────────────────────────────────────────────────────────────────────────
-
-[16:20:16.124] [RESULT] --- Zombie Session Evidence ---
-[16:20:16.124] [RESULT]   session.closed:    false    (should be true, but isn't)
-[16:20:16.124] [RESULT]   session.destroyed: false (should be true, but isn't)
-[16:20:16.124] [RESULT]   responses:         0/5    (requests vanished into the black hole)
-[16:20:16.124] [RESULT]   events received:   NONE  (no close, no error)
-[16:20:16.125] [RESULT]   ping callback:     never fired
-[16:20:16.125] [RESULT]   outboundQueueSize: 0    (0 is normal — small requests fit in kernel TCP buffer)
-[16:20:16.125] [RESULT] 
-[16:20:16.125] [RESULT]   The session is a ZOMBIE: the OS connection is broken,
-[16:20:16.125] [RESULT]   but Node.js has no idea. All APIs report "healthy".
-
-[16:20:16.125] [PHASE 6] !!! TRIGGERING ASSERTION CRASH !!!
-[16:20:16.125] [PHASE 6] Writing directly to TLS socket to create concurrent write...
-[16:20:16.125] [PHASE 6] This will crash with: Assertion failed: !current_write_
-[16:20:16.125] [PHASE 6]   at node::crypto::TLSWrap::DoWrite (src/crypto/crypto_tls.cc)
-
-[16:20:16.126] [PHASE 6]   TLS socket: destroyed=false writable=true
-[16:20:16.126] [PHASE 6]   Writing raw data to TLS socket NOW...
-
-  #  node[10210]: virtual void node::http2::Http2Session::OnStreamAfterWrite(WriteWrap *, int) at ../src/node_http2.cc:1774
-  #  Assertion failed: is_write_in_progress()
-
------ Native stack trace -----
-
- 1: 0x1004d2bb0 node::Assert(node::AssertionInfo const&) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 2: 0x10233e788 node::http2::Http2Session::OnStreamAfterWrite(node::WriteWrap*, int) (.cold.1) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 3: 0x1005060d0 node::http2::Http2Session::MaybeNotifyGracefulCloseComplete() [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 4: 0x1005ec8a8 node::WriteWrap::OnDone(int) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 5: 0x1005ecbd0 node::StreamReq::Done(int, char const*) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 6: 0x100685fdc node::crypto::TLSWrap::InvokeQueued(int, char const*) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 7: 0x100687d7c node::crypto::TLSWrap::OnStreamAfterWrite(node::WriteWrap*, int) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 8: 0x100460878 node::Environment::RunAndClearNativeImmediates(bool)::$_0::operator()(node::CallbackQueue<void, node::Environment*>*) const [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
- 9: 0x10045fbc8 node::Environment::RunAndClearNativeImmediates(bool) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-10: 0x10045f174 node::Environment::CheckImmediate(uv_check_s*) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-11: 0x1010c20b4 uv__run_check [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-12: 0x1010bb0d4 uv_run [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-13: 0x1003d2518 node::SpinEventLoopInternal(node::Environment*) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-14: 0x10051d93c node::NodeMainInstance::Run() [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-15: 0x10048dd50 node::Start(int, char**) [/Users/jesuswahrman/.nvm/versions/node/v22.22.0/bin/node]
-16: 0x187577da4 start [/usr/lib/dyld]
