+++ x-pack/platform/plugins/shared/fleet/moon.yml
@@ -124,6 +124,7 @@ dependsOn:
   - '@kbn/rison'
   - '@kbn/rule-data-utils'
   - '@kbn/doc-links'
+  - '@kbn/yaml-loader'
 tags:
   - plugin
   - prod
