+++ x-pack/solutions/security/plugins/security_solution/moon.yml
@@ -48,6 +48,7 @@ dependsOn:
   - '@kbn/spaces-plugin'
   - '@kbn/timelines-plugin'
   - '@kbn/es-query'
+  - '@kbn/as-code-shared-transforms'
   - '@kbn/utility-types'
   - '@kbn/securitysolution-list-utils'
   - '@kbn/securitysolution-io-ts-list-types'
